+++ Basic/Invocation-with-odd-paths/test.2.dot-sourcing.ps1
@@ -9,7 +9,7 @@ Set-Location -LiteralPath $PSScriptRoot
 $null = mkdir test[ -Force
 Set-Content -LiteralPath test[\test.ps1 -Value 42
 
-# this works in v2, v3, v4 and fails in v5
+# this works in v2, v3, v4, v7 and fails in v5, v6
 try {
 	. '.\test[\test.ps1'
 }
@@ -18,6 +18,7 @@ catch {
 }
 
 # this fails due to the invalid wildcard
+# fixed in v7
 try {
 	. "$PSScriptRoot\test[\test.ps1"
 }