+++ Basic/Provider-specific-Filter/Test-2.1.FileSystem.ps1
@@ -4,7 +4,7 @@ Set-Location -LiteralPath (Split-Path $MyInvocation.MyCommand.Definition)
 # file name with no dots
 Set-Content tmp ''
 
-# Filter *.* includes "tmp"
+# Filter *.* includes "tmp" in v2-v5, v7. But v6 excludes "tmp"
 $r = Get-ChildItem . -Filter *.*
 $r | Select-Object -ExpandProperty Name