+++ Cmdlets/Group-Object/Expression-with-no-value/Test-1.ps1
@@ -5,6 +5,7 @@ $x = 1 | Select-Object -Property @{Name = 'X'; Expression = {}}
 # Show that the property X exists
 $x | Get-Member X
 
-# v5: fails
-# v2: works
+# v7 works
+# v5, v6 fail
+# v2 works
 $x | Group-Object -Property X